feat: change reward type; int -> double
프론트 요청에 따라 탄소발자국 리워드 타입 변경

+++ src/main/java/com/solution/green/dto/MemberDto.java
@@ -46,7 +46,7 @@ public class MemberDto {
         @NotNull
         private String password;
         @NotNull
-        private Integer reward;
+        private Double reward;
         @Nullable
         private String imageUrl;
         @Setter
